+++ CHANGELOG.md
@@ -9,3 +9,4 @@
     - `burst` snaps to vertices
     - `line` snaps to midpoints or vertices
 - Force `gridTemplates` on for hex grids
+- Snap Circle and Cone to 30 degree angle increments